First part of cleanup in detectors.
