+++ src/filters/drop.cpp
@@ -2,188 +2,10 @@
 
 #include "filters/drop.hpp"
 
-#include "messages/serenity.hpp"
-
 #include "serenity/wid.hpp"
-
-#include "stout/none.hpp"
 
 namespace mesos {
 namespace serenity {
-
-Result<ChangePointDetection> NaiveChangePointDetector::processSample(
-    double_t in) {
-  if (this->contentionCooldownCounter > 0) {
-    this->contentionCooldownCounter--;
-
-    return None();
-  }
-
-  if (in < this->state.absoluteThreshold) {
-    this->contentionCooldownCounter = this->state.contentionCooldown;
-    ChangePointDetection cpd;
-    cpd.severity = this->state.absoluteThreshold - in;
-    return cpd;
-  }
-
-  return None();
-}
-
-
-Result<ChangePointDetection> RollingChangePointDetector::processSample(
-    double_t in) {
-  this->window.push_back(in);
-
-  if (this->window.size() < this->state.windowSize) {
-    return None();  // Only warm up.
-  }
-
-  double_t basePoint = this->window.front();
-  this->window.pop_front();
-
-  if (this->contentionCooldownCounter > 0) {
-    this->contentionCooldownCounter--;
-    return None();
-  }
-
-  if (in < (basePoint - this->state.relativeThreshold)) {
-    this->contentionCooldownCounter = this->state.contentionCooldown;
-    ChangePointDetection cpd;
-
-    // We calculate severity as difference between threshold and actual
-    // processed value.
-    cpd.severity = (basePoint - this->state.relativeThreshold) - in;
-    return cpd;
-  }
-
-  if (in < basePoint) {
-    SERENITY_LOG(INFO) << " Found decrease, "
-                 "but not significant: " << (in - basePoint);
-  }
-
-  return None();
-}
-
-
-Result<ChangePointDetection> RollingFractionalDetector::processSample(
-    double_t in) {
-  this->window.push_back(in);
-
-  if (this->window.size() < this->state.windowSize) {
-    return None();  // Only warm up.
-  }
-
-  double_t basePoint = this->window.front();
-  this->window.pop_front();
-
-  if (this->contentionCooldownCounter > 0) {
-    this->contentionCooldownCounter--;
-    return None();
-  }
-
-  // Current drop fraction indicates how much value has drop in relation to
-  // base point
-  double_t currentDropFraction = 1.0 - (in / basePoint);
-
-  SERENITY_LOG(INFO)
-            << " inputValue: " << in
-            << " | baseValue = " << basePoint
-            << " | current drop %: " << currentDropFraction * 100
-            << " | threshold %: " << this->state.fractionalThreshold * 100;
-
-  // If drop fraction is higher than threshold, then trigger contention.
-  if (currentDropFraction > this->state.fractionalThreshold) {
-    this->contentionCooldownCounter = this->state.contentionCooldown;
-    ChangePointDetection cpd;
-
-    cpd.severity = currentDropFraction * this->state.severityLevel;
-    LOG(INFO) << tag.NAME() << " Created contention with severity = "
-              << cpd.severity;
-    return cpd;
-  }
-
-  if (in < basePoint) {
-    SERENITY_LOG(INFO) << " Found decrease, "
-        "but not significant: " << (currentDropFraction * 100) << "%";
-  }
-
-  return None();
-}
-
-Result<ChangePointDetection> AssuranceFractionalDetector::processSample(
-    double_t in) {
-  this->window.push_back(in);
-
-  if (this->window.size() < this->state.windowSize) {
-    return None();  // Only warm up.
-  }
-
-  double_t basePoint = this->window.front();
-  this->window.pop_front();
-
-  if (this->referencePoint.isSome()) {
-    // Check if the signal returned to normal state. (!)
-    double_t nearValue = this->state.nearFraction * this->referencePoint.get();
-    SERENITY_LOG(INFO) << "Waiting for signal to return to "
-                       << (this->referencePoint.get() - nearValue)
-                       << " (base value) after "
-                       << "corrections. Waiting iteration: "
-                       << referencePointCounter;
-    this->referencePointCounter++;
-    // We want to use reference Base Point instead of base point.
-    basePoint = in;
-    if (in >= (this->referencePoint.get() - nearValue)) {
-      this->referencePoint = None();
-      SERENITY_LOG(INFO) << "Signal returned to established state.";
-    }
-  }
-
-  if (this->contentionCooldownCounter > 0) {
-    SERENITY_LOG(INFO) << "Cooldown for: " << contentionCooldownCounter;
-    this->contentionCooldownCounter--;
-    return None();
-  }
-
-  // Current drop fraction indicates how much value has drop in relation to
-  // base point
-  double_t currentDropFraction = 1.0 - (in / basePoint);
-
-  SERENITY_LOG(INFO)
-  << " inputValue: " << in
-  << " | baseValue = " << basePoint
-  << " | current drop %: " << currentDropFraction * 100
-  << " | threshold %: " << this->state.fractionalThreshold * 100;
-
-  // If drop fraction is higher than threshold or signal did not returned
-  // after cooldown, then trigger contention.
-  if (currentDropFraction > this->state.fractionalThreshold ||
-      (this->referencePoint.isSome())) {
-    ChangePointDetection cpd;
-    if (this->referencePoint.isNone()) {
-      cpd.severity =  currentDropFraction * this->state.severityLevel;
-      this->lastSeverity = cpd.severity;
-    } else {
-      cpd.severity =  this->lastSeverity;
-    }
-
-    LOG(INFO) << tag.NAME() << " Created contention with severity = "
-              << cpd.severity;
-
-    this->contentionCooldownCounter = this->state.contentionCooldown;
-    this->referencePoint = basePoint;
-    this->referencePointCounter = 0;
-
-    return cpd;
-  }
-
-  if (in < basePoint) {
-    SERENITY_LOG(INFO) << " Found decrease, "
-        "but not significant: " << (currentDropFraction * 100) << "%";
-  }
-
-  return None();
-}
-
 
 template <typename T>
 Try<Nothing> DropFilter<T>::consume(const ResourceUsage& in) {
@@ -254,14 +76,6 @@ Try<Nothing> DropFilter<T>::consume(const ResourceUsage& in) {
   return Nothing();
 }
 
-
-// Fix for using templated methods in .cpp file.
-// See:
-//    https://isocpp.org/wiki/faq/templates#separate-template-fn-defn-from-decl
-template class DropFilter<NaiveChangePointDetector>;
-template class DropFilter<RollingChangePointDetector>;
-template class DropFilter<RollingFractionalDetector>;
-template class DropFilter<AssuranceFractionalDetector>;
 
 }  // namespace serenity
 }  // namespace mesos
