+++ src/serenity/default_vars.hpp
@@ -1,6 +1,9 @@
 #ifndef SERENITY_DEFAULT_VARS_HPP
 #define SERENITY_DEFAULT_VARS_HPP
 
+#include <cstdint>
+#include <cmath>
+
 namespace mesos {
 namespace serenity {
 
@@ -20,18 +23,18 @@ namespace changepoint {
 
 constexpr uint64_t DEFAULT_WINDOW_SIZE = 10;
 constexpr uint64_t DEFAULT_CONTENTION_COOLDOWN = 10;
-constexpr double DEFAULT_ABS_THRESHOLD = 0;
-constexpr double DEFAULT_RELATIVE_THRESHOLD = 20;
-constexpr double DEFAULT_FRACTIONAL_THRESHOLD = 0.5;
-constexpr double DEFAULT_SEVERITY_FRACTION = 0.4;
-constexpr double DEFAULT_NEAR_FRACTION = 0.1;
+constexpr double_t DEFAULT_ABS_THRESHOLD = 0;
+constexpr double_t DEFAULT_RELATIVE_THRESHOLD = 20;
+constexpr double_t DEFAULT_FRACTIONAL_THRESHOLD = 0.5;
+constexpr double_t DEFAULT_SEVERITY_FRACTION = 0.4;
+constexpr double_t DEFAULT_NEAR_FRACTION = 0.1;
 
 }  // namespace changepoint
 
 
 namespace utilization {
 
-constexpr double DEFAULT_THRESHOLD = 0.95;
+constexpr double_t DEFAULT_THRESHOLD = 0.95;
 
 }  // namespace utilization
